# src/UI/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/expenses": {
      "get": {
        "tags": [
          "Api"
        ],
        "operationId": "Get Expenses",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Expense"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Expense"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Expense"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Api"
        ],
        "operationId": "Create Expense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
    "/api/expenses/{expenseId}": {
      "put": {
        "tags": [
          "Api"
        ],
        "operationId": "Update Expense",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Api"
        ],
        "operationId": "Delete Expense",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          }
        }
      }
    },
    "/api/expense-categories": {
      "get": {
        "tags": [
          "Api"
        ],
        "operationId": "Get Expense Categories",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseCategory"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseCategory"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseCategory"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Api"
        ],
        "operationId": "Create Expense Category",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCategory"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCategory"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseCategory"
                }
              }
            }
          }
        }
      }
    },
    "/api/expense-categories/{expenseCategoryId}": {
      "put": {
        "tags": [
          "Api"
        ],
        "operationId": "Update Expense Category",
        "parameters": [
          {
            "name": "expenseCategoryId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCategory"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCategory"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseCategory"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Api"
        ],
        "operationId": "Delete Expense Category",
        "parameters": [
          {
            "name": "expenseCategoryId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseCategory"
                }
              }
            }
          }
        }
      }
    },
    "/api/incomes": {
      "get": {
        "tags": [
          "Api"
        ],
        "operationId": "Get Incomes",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Income"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Income"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Income"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Api"
        ],
        "operationId": "Create Income",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Income"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Income"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Income"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              }
            }
          }
        }
      }
    },
    "/api/incomes/{incomeId}": {
      "put": {
        "tags": [
          "Api"
        ],
        "operationId": "Update Income",
        "parameters": [
          {
            "name": "incomeId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Income"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Income"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Income"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Api"
        ],
        "operationId": "Delete Income",
        "parameters": [
          {
            "name": "incomeId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              }
            }
          }
        }
      }
    },
    "/api/income-categories": {
      "get": {
        "tags": [
          "Api"
        ],
        "operationId": "Get Income Categories",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeCategory"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeCategory"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeCategory"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Api"
        ],
        "operationId": "Create Income Category",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeCategory"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeCategory"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeCategory"
                }
              }
            }
          }
        }
      }
    },
    "/api/income-categories/{incomeCategoryId}": {
      "put": {
        "tags": [
          "Api"
        ],
        "operationId": "Update Income Category",
        "parameters": [
          {
            "name": "incomeCategoryId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeCategory"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeCategory"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeCategory"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Api"
        ],
        "operationId": "Delete Income Category",
        "parameters": [
          {
            "name": "incomeCategoryId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeCategory"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Expense": {
        "type": "object",
        "properties": {
          "expenseID": {
            "type": "integer",
            "format": "int32"
          },
          "expenseCategoryID": {
            "type": "integer",
            "format": "int32"
          },
          "vendorID": {
            "type": "integer",
            "format": "int32"
          },
          "paymentMethodID": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "date": {
            "type": "String",
            "format": "date"
          },
          "note": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "expenseCategory": {
            "$ref": "#/components/schemas/ExpenseCategory"
          },
          "vendor": {
            "$ref": "#/components/schemas/Vendor"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "additionalProperties": false
      },
      "ExpenseCategory": {
        "type": "object",
        "properties": {
          "expenseCategoryID": {
            "type": "integer",
            "format": "int32"
          },
          "displayName": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Income": {
        "type": "object",
        "properties": {
          "incomeID": {
            "type": "integer",
            "format": "int32"
          },
          "incomeCategoryID": {
            "type": "integer",
            "format": "int32"
          },
          "incomeSourceID": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "date": {
            "type": "String",
            "format": "date"
          },
          "note": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "incomeCategory": {
            "$ref": "#/components/schemas/IncomeCategory"
          },
          "incomeSource": {
            "$ref": "#/components/schemas/IncomeSource"
          }
        },
        "additionalProperties": false
      },
      "IncomeCategory": {
        "type": "object",
        "properties": {
          "incomeCategoryID": {
            "type": "integer",
            "format": "int32"
          },
          "displayName": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}